refactor: clean up best hit computation
